voyager/commands/init.py
#!/usr/bin/env python3

import os
import sys
import click
from pathlib import Path
import yaml

from ..utils import check_git_repo, get_repo_info

@click.command('init')
@click.option('--concourse-url', help='Concourse CI API URL')
@click.option('--concourse-team', help='Concourse CI team name')
@click.option('--pipeline', help='Concourse pipeline name')
def init_repo(concourse_url, concourse_team, pipeline):
    """Initialize the current repository for Voyager."""
    if not check_git_repo():
        click.echo("Error: Current directory is not a git repository", err=True)
        sys.exit(1)
    
    try:
        owner, repo = get_repo_info()
        click.echo(f"Initializing Voyager for {owner}/{repo}...")
        
        # Full implementation would go here
        click.echo("Init command not fully implemented yet.")
    
    except Exception as e:
        click.echo(f"Error initializing repository: {str(e)}", err=True)
        sys.exit(1)
