requirements.txt
# Requirements for Voyager
click>=8.0.0         # Command-line interface creation
gitpython>=3.1.0     # Git repository interaction
requests>=2.25.0     # HTTP API interactions
semver>=2.13.0       # Semantic versioning
pyyaml>=6.0          # YAML file handling
